+++ bank-feeds/src/codatbankfeeds/models/operations/update_source_account.py
@@ -12,9 +12,11 @@ from typing import Optional
 @dataclasses.dataclass
 class UpdateSourceAccountRequest:
     account_id: str = dataclasses.field(metadata={'path_param': { 'field_name': 'accountId', 'style': 'simple', 'explode': False }})
-    r"""Unique identifier for an account"""
+    r"""Unique identifier for an account."""
     company_id: str = dataclasses.field(metadata={'path_param': { 'field_name': 'companyId', 'style': 'simple', 'explode': False }})
+    r"""Unique identifier for a company."""
     connection_id: str = dataclasses.field(metadata={'path_param': { 'field_name': 'connectionId', 'style': 'simple', 'explode': False }})
+    r"""Unique identifier for a connection."""
     source_account: Optional[shared_sourceaccount.SourceAccount] = dataclasses.field(default=None, metadata={'request': { 'media_type': 'application/json' }})
     
 
@@ -24,10 +26,13 @@ class UpdateSourceAccountRequest:
 @dataclasses.dataclass
 class UpdateSourceAccountResponse:
     content_type: str = dataclasses.field()
+    r"""HTTP response content type for this operation"""
     status_code: int = dataclasses.field()
+    r"""HTTP response status code for this operation"""
     error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
     r"""Your API request was not properly authorized."""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
+    r"""Raw HTTP response; suitable for custom response parsing"""
     source_account: Optional[shared_sourceaccount.SourceAccount] = dataclasses.field(default=None)
     r"""Success"""